+++ tests/testthat/tests-specifics.R
@@ -1,9 +1,0 @@
-# OLS
-# OLS in glm
-# OLS interactions
-# logit
-# probit
-# poisson
-# glm interactions
-# factors
-# factor by continuous interaction