+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## margins 0.2.5
 
+* Added better support for factor `x` variables in `cplot()`.
 * Added (rudimentary) tests of variance methods. (#21)
 * Removed `.build_predict_fun()` factory function, as it was no longer needed.
 * Fix vignettes so package can be built with them. (#16)